pyproject.toml
[tool.poetry]
name = "douyin-tiktok-download-api"
version = "0.1.0"
description = "Douyin/TikTok/Bilibili async scraping API and web app"
authors = ["Unknown <unknown@example.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
  { include = "app" },
  { include = "crawlers" }
]

[tool.poetry.dependencies]
# Support Python 3.11/3.12 to avoid building native deps on 3.13
python = ">=3.11,<3.13"
aiofiles = "23.2.1"
annotated-types = "0.6.0"
anyio = "4.3.0"
browser-cookie3 = "0.19.1"
certifi = "2024.2.2"
click = "8.1.7"
colorama = "0.4.6"
fastapi = "0.110.2"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
idna = "3.7"
importlib-resources = "6.4.0"
lz4 = "4.3.3"
markdown-it-py = "3.0.0"
mdurl = "0.1.2"
numpy = "*"
pycryptodomex = "3.20.0"
pydantic = "2.7.0"
pydantic-core = "2.18.1"
pyfiglet = "1.0.0"
pygments = "2.17.2"
pypng = "0.20220715.0"
pywebio = "1.8.3"
pywebio-battery = "0.6.0"
PyYAML = "6.0.1"
qrcode = "7.4.2"
rich = "13.7.1"
sniffio = "1.3.1"
starlette = "0.37.2"
tornado = "6.4"
typing-extensions = "4.11.0"
ua-parser = "0.18.0"
user-agents = "2.2.0"
uvicorn = "0.29.0"
websockets = "12.0"
gmssl = "3.2.2"
tenacity = "~9.0.0"
pyqt6 = "^6.9.1"
qtawesome = "^1.4.0"
pyinstaller = "^6.15.0"
pillow = "^11.3.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"


